third_party/blink/web_tests/external/wpt/css/css-conditional/container-queries/style-query-no-cycle-expected.txt

This is a testharness.js-based test.
[FAIL] style(--foo: var(--foo)) should match
  assert_equals: expected "rgb(0, 128, 0)" but got "rgb(0, 0, 0)"
Harness: the test ran to completion.
